% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3read10X.R
\name{s3.read10X}
\alias{s3.read10X}
\title{Read in 10x output files from an AWS S3 bucket and create a Seurat Object}
\usage{
s3.read10X(bucket, region, project = NULL, min.cells = 0, min.features = 0)
}
\arguments{
\item{bucket}{Scalar character, the AWS S3 bucket containing the Cell Ranger
outputs (matrix.mtx.gz, features.tsv.gz, and barcodes.tsv.gz)}

\item{region}{Sclalar character, the region for the S3 bucket (usually
"us-east-1" for raw data files)}

\item{project}{Scalar character or \code{NULL}.}

\item{min.cells}{Scalar count, the minimum number of cells a feature must be
detected in to be retained.}

\item{min.features}{Scalar count, the minimum number of features detected in
a cell to be retained.}
}
\value{
Returns a Seurat object
}
\description{
Read in 10x output files from an AWS S3 bucket and create a Seurat Object
}
