% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleVolcano.R
\name{simpleVolcano}
\alias{simpleVolcano}
\title{Plot a volcano plot both static and with plotly and produce a DT::datatable}
\usage{
simpleVolcano(
  DE,
  fcColumn,
  pColumn,
  geneCol,
  fcMin = 0.5,
  pMax = 0.05,
  labels = NULL
)
}
\arguments{
\item{DE}{\itemize{
\item data frame with differential expression results
(eg: toptable output, DESeq2::results ouput, etc.)
}}

\item{fcColumn}{scalar character, the column with logFoldChanges in \code{DE}.}

\item{pColumn}{scalar character, the column with pvalues in \code{DE}.}

\item{geneCol}{scalar character, the column with gene names in \code{DE}.}

\item{fcMin}{Scalar numeric, the absolute FC cutoff to be used
(defaults to 0.5)}

\item{pMax}{calar numeric, The p-value cutoff to be used (defaults to 0.05)}

\item{labels}{Character vector of genes to be plotted, elements must match
the \code{geneCol} of \code{DE}.}
}
\value{
A ggplot object
}
\description{
Plot a volcano plot both static and with plotly and produce a DT::datatable
}
\examples{
\dontrun{
volcanoPlots(DE, fcColumn, pColumn, geneCol)
}
}
